feat: Implement web routes, permission middleware, and service provider for user, role, and lead management, including a new leads index view.

+++ Modules/CRM/Lead/routes/web.php
@@ -7,10 +7,10 @@ Route::prefix('admin/leads')
     ->middleware(['web', 'auth'])
     ->name('admin.leads.')
     ->group(function () {
-        Route::get('/', [LeadWebController::class, 'index'])->name('index');
-        Route::get('/create', [LeadWebController::class, 'create'])->name('create');
-        Route::post('/', [LeadWebController::class, 'store'])->name('store');
-        Route::get('/{id}/edit', [LeadWebController::class, 'edit'])->name('edit');
-        Route::put('/{id}', [LeadWebController::class, 'update'])->name('update');
-        Route::delete('/{id}', [LeadWebController::class, 'destroy'])->name('destroy');
+        Route::get('/', [LeadWebController::class, 'index'])->name('index')->middleware('permission:leads.view');
+        Route::get('/create', [LeadWebController::class, 'create'])->name('create')->middleware('permission:leads.create');
+        Route::post('/', [LeadWebController::class, 'store'])->name('store')->middleware('permission:leads.create');
+        Route::get('/{id}/edit', [LeadWebController::class, 'edit'])->name('edit')->middleware('permission:leads.update');
+        Route::put('/{id}', [LeadWebController::class, 'update'])->name('update')->middleware('permission:leads.update');
+        Route::delete('/{id}', [LeadWebController::class, 'destroy'])->name('destroy')->middleware('permission:leads.delete');
     });
